sol.video.requirements.txt
gpiozero
- fastapi
- Jinja2
pyglet
matplotlib
numpy
opencv-contrib-python
pillow
requests
ruamel.yaml
scipy
urllib3
uvicorn
